requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
distro==1.9.0
dnspython==2.5.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
isodate==0.6.1
jsonschema==4.21.1
jsonschema-path==0.3.2
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.10.0
MarkupSafe==2.1.4
more-itertools==10.2.0
motor==3.3.2
multidict==6.0.4
numpy==1.26.3
openai==1.9.0
openapi-core==0.18.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
packaging==23.2
parse==1.20.0
pathable==0.4.3
prance==23.6.21.0
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.1
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rpds-py==0.17.1
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
semantic-kernel==0.4.6.dev0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
yarl==1.9.4
